pyproject.toml
[tool.poetry]
name = "coursework-7"
version = "0.1.0"
description = ""
authors = ["Ильдар Галеев <ildar-galeev92@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
pytelegrambotapi = "^4.16.1"
environs = {extras = ["django"], version = "^10.0.0"}
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.3.6"
eventlet = "^0.35.2"
redis = "^5.0.3"
django-celery-beat = "^2.6.0"
django-filter = "^24.1"
django-timezone-field = "^6.1.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ipython = "^8.22.2"
coverage = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
